Prototype A: podman-remote
This version uses the Pbench ARCA plugin to drive a sidecar container running
Pbench Agent code. Each "step" models a state of the Tool Meister ecosystem
from "initialize" through "start", "stop", "finalize", and "upload".

It lacks a mechanism to propagate context from one "step" to the next, which
means that only "initialize" actually works. I'm abandoning this approach in
favor of a single "step" with an externally controlled state machine to drive
it through the same states: this will fit nicely with the "signals" model
proposed for ARCA.